+++ app/src/main/java/com/example/weatherapplication/domain/model/current/MetricTemperature.kt
@@ -1,6 +1,6 @@
 package com.example.weatherapplication.domain.model.current
 
 data class MetricTemperature(
-    override val temp: Double,
-    override val feelsLike: Double
+    override val temp: Int,
+    override val feelsLike: Int
 ) : Temperature
