+++ app/src/main/java/com/example/weatherapplication/domain/model/day/DayTemperature.kt
@@ -1,7 +1,7 @@
 package com.example.weatherapplication.domain.model.day
 
 sealed interface DayTemperature {
-    val maxTemp: Double
-    val minTemp: Double
-    val avgTemp: Double
+    val maxTemp: Int
+    val minTemp: Int
+    val avgTemp: Int
 }
